pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "openai>=1.101.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "slack-sdk>=3.36.0",
]
